fix: add default network provider
