+++ src/main/java/com/aws/greengrass/clientdevices/auth/ClientDevicesAuthService.java
@@ -19,7 +19,7 @@ import com.aws.greengrass.clientdevices.auth.configuration.GroupConfiguration;
 import com.aws.greengrass.clientdevices.auth.configuration.GroupManager;
 import com.aws.greengrass.clientdevices.auth.configuration.RuntimeConfiguration;
 import com.aws.greengrass.clientdevices.auth.connectivity.CISShadowMonitor;
-import com.aws.greengrass.clientdevices.auth.infra.NetworkState;
+import com.aws.greengrass.clientdevices.auth.infra.NetworkStateProvider;
 import com.aws.greengrass.clientdevices.auth.session.MqttSessionFactory;
 import com.aws.greengrass.clientdevices.auth.session.SessionConfig;
 import com.aws.greengrass.clientdevices.auth.session.SessionCreator;
@@ -112,7 +112,7 @@ public class ClientDevicesAuthService extends PluginService {
     private void initializeInfrastructure() {
         // Infra setup
         context.put(RuntimeConfiguration.class, RuntimeConfiguration.from(getRuntimeConfig()));
-        NetworkState networkState = context.get(NetworkState.class);
+        NetworkStateProvider networkState = context.get(NetworkStateProvider.class);
         networkState.registerHandler(context.get(CISShadowMonitor.class));
         networkState.registerHandler(context.get(BackgroundCertificateRefresh.class));
         context.get(BackgroundCertificateRefresh.class).start();
