+++ src/main/java/com/aws/greengrass/clientdevices/auth/iot/usecases/VerifyThingAttachedToCertificate.java
@@ -7,7 +7,6 @@ package com.aws.greengrass.clientdevices.auth.iot.usecases;
 
 import com.aws.greengrass.clientdevices.auth.api.UseCases;
 import com.aws.greengrass.clientdevices.auth.exception.CloudServiceInteractionException;
-import com.aws.greengrass.clientdevices.auth.infra.NetworkState;
 import com.aws.greengrass.clientdevices.auth.infra.NetworkStateProvider;
 import com.aws.greengrass.clientdevices.auth.iot.IotAuthClient;
 import com.aws.greengrass.clientdevices.auth.iot.Thing;
@@ -23,7 +22,7 @@ import javax.inject.Inject;
 public class VerifyThingAttachedToCertificate
         implements UseCases.UseCase<Boolean, VerifyThingAttachedToCertificateDTO> {
     private final IotAuthClient iotAuthClient;
-    private final NetworkState networkState;
+    private final NetworkStateProvider networkState;
     private final ThingRegistry thingRegistry;
     private static final Logger logger = LogManager.getLogger(VerifyThingAttachedToCertificate.class);
 
@@ -38,7 +37,7 @@ public class VerifyThingAttachedToCertificate
      */
     @Inject
     public VerifyThingAttachedToCertificate(IotAuthClient iotAuthClient, ThingRegistry thingRegistry,
-                                  NetworkState networkState) {
+                                            NetworkStateProvider networkState) {
         this.iotAuthClient = iotAuthClient;
         this.thingRegistry = thingRegistry;
         this.networkState = networkState;
